requirements.txt
streamlit
transformers
torch
sentence-transformers
faiss-cpu
numpy